feat(agents): Pass C — Workspace tabs + Agents management grid
Split the AgentsScreen monolith into two surfaces:
- Workspace nav (was "Chat", new PanelsTopLeft icon) hosts a Chat | Agent
  tab bar. Chat = plain LLM chat; Agent = agent conversations — the chat
  half moved verbatim into workspace/AgentChatView.
- Agents nav is now pure management: a grid of agent cards → full-page
  edit (agents/AgentEditPage). No conversations there.

Chat-tab sidebar now excludes agent conversations. Routes moved to
/workspace/{chat,agent}; /chat/:convId kept so LAN share links survive.
