+++ turbollm/web/src/screens/chat/ConversationSidebar.tsx
@@ -74,7 +74,8 @@ export function ConversationSidebar({
   const searchRef = useRef<HTMLInputElement>(null)
   const mut = useConversationMutations()
   const convsQ = useConversations(debouncedQ || undefined)
-  const convs = convsQ.data?.conversations ?? []
+  // Agent conversations live in Workspace → Agent; the Chat tab shows plain chats only.
+  const convs = (convsQ.data?.conversations ?? []).filter((c) => !c.agentId)
   const foldersQ = useFolders()
   const folders = foldersQ.data?.folders ?? []
